@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { MatDialog, MatDialogRef } from '@angular/material';
+import { Observable } from 'rxjs';
 
-import { Observable } from 'rxjs/Observable';
 import { DeleteDialogComponent } from './delete-dialog.component';
 
 @Injectable()
